+++ web/modules/custom/giv_din_stemme/src/Form/GivDinStemmeSettingsForm.php
@@ -51,14 +51,21 @@ class GivDinStemmeSettingsForm extends FormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
-    try {
-      $permissionsHelpPageNode = $this->entityTypeManager
-        ->getStorage('node')
-        ->load(
-          $this->state->get('giv_din_stemme.permissions_help_page')
-        );
-    }
-    catch (InvalidPluginDefinitionException | PluginNotFoundException $e) {
+    $references = [];
+    $storage = $this->entityTypeManager->getStorage('node');
+    foreach ([
+      'permissions_help_page',
+      'safari_on_ios_help_page',
+    ] as $key) {
+      try {
+        if ($id = $this->state->get('giv_din_stemme.' . $key)) {
+          if ($node = $storage->load($id)) {
+            $references[$key] = $node;
+          }
+        }
+      }
+      catch (InvalidPluginDefinitionException | PluginNotFoundException $e) {
+      }
     }
 
     $form['general_settings'] = [
@@ -126,7 +133,15 @@ class GivDinStemmeSettingsForm extends FormBase {
       '#type' => 'entity_autocomplete',
       '#target_type' => 'node',
       '#title' => $this->t('Permissions help page'),
-      '#default_value' => $permissionsHelpPageNode ?? NULL,
+      '#default_value' => $references['permissions_help_page'] ?? NULL,
+      '#weight' => '0',
+    ];
+
+    $form['references']['safari_on_ios_help_page'] = [
+      '#type' => 'entity_autocomplete',
+      '#target_type' => 'node',
+      '#title' => $this->t('Safari on iOS help page'),
+      '#default_value' => $references['safari_on_ios_help_page'] ?? NULL,
       '#weight' => '0',
     ];
 
@@ -151,6 +166,7 @@ class GivDinStemmeSettingsForm extends FormBase {
       'giv_din_stemme.donate_page_text' => $form_state->getValue('donate_page_text')['value'],
       'giv_din_stemme.thank_you_text' => $form_state->getValue('thank_you_text')['value'],
       'giv_din_stemme.permissions_help_page' => $form_state->getValue('permissions_help_page'),
+      'giv_din_stemme.safari_on_ios_help_page' => $form_state->getValue('safari_on_ios_help_page'),
     ]);
 
     drupal_flush_all_caches();